Restructure internal/pipelines directory
