+++ internal/pipeline/config_test.go
@@ -1,4 +1,4 @@
-package pipelines
+package pipeline
 
 import (
 	"os"
