+++ internal/pipeline/path.go
@@ -1,4 +1,4 @@
-package pipelines
+package pipeline
 
 import (
 	"strings"
@@ -8,7 +8,6 @@ import (
 )
 
 func ResolveFromPath(path string, org string, client *buildkite.Client) ([]string, error) {
-
 	repos, err := getRepoURLs(path)
 	if err != nil {
 		return nil, err
@@ -17,7 +16,6 @@ func ResolveFromPath(path string, org string, client *buildkite.Client) ([]strin
 }
 
 func filterPipelines(repoURLs []string, org string, client *buildkite.Client) ([]string, error) {
-
 	var currentPipelines []string
 	page := 1
 	per_page := 30
