+++ internal/pipeline/pipeline.go
@@ -1,0 +1,7 @@
+package pipeline
+
+// Pipeline is a struct containing information about a pipeline for a resolver to return
+type Pipeline struct {
+	Name string
+	Org  string
+}
